requirements.txt
# requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-groq==0.0.1
sqlalchemy==2.0.25
mysql-connector-python==8.2.0
streamlit==1.30.0
python-decouple==3.8